Use /67 channel controller bits in device address for TCH instruction
The 2067-2 had an EC mode feature not documented in the Functional
Characteristics to select the channel controller to use for testing
channel existence.  This is so the CPU could figure out all possible
paths to a device through its channel controllers for configuring its
I/O path management.

The feature applied to the TCH and, possibly, the TIO instructions.
TCH allows channel connections through CC 0 only.  TIO does not
parse the CC field of the device address and may lead to unexpected
behaviour if used.