IBM360: Implement 2703 BSC protocol for HASP RJE use.
These changes implement IBM's BSC (bisync) point-to-point
protocol for use with devices that connect to a CPU (through
the 2703) via dialup or leased telephone lines to submit and
retrieve batch jobs remotely.  On IBM 360 and 370 systems, HASP
was the program that provided RJE (remote job entry) services
and was the principal agent for BSC communications.

BSC is implemented by stealing a small number of lines from the
COM/COML device (configured as NUM_UNITS_BSC in ibm360_defs.h)
and treating them as BSC lines.  (This mimics having bisync SF
adaptors on 2703s for leased RJE lines.)  These units constitute
a BSC group to which TCP connections can be made to communicate
with the device.  These device addresses lie at the end of a
COML unit group in order to share the service and I/O initiation
routines common to the 2703.  They have a separate TCP port from
the COML units to separate the BSC functionality.  The BSC
devices are defined in the simh config file using a line like,

attach bsc 2780,line=0

which declares the TCP port (here 2780) for RJE connections and
enables use of the BSC protocol on those lines.  NUM_DEVS_BSC in
ibm360_defs.h controls the max number of bsc devices that may be
configured.

BSC support uses the simh multiplexer routines to manage TCP
connections.  Unfortunately, the protocol involves disconnects
(via the DISABLE/ENABLE CCWs) to stop and restart different
transaction modes.  The multiplexer routines do not seem to be
robust to these and there is always a failure to reconnect
after a previous disconnect.  A retry of the reconnection recovers
the TCP functionality, but is irritating from a user viewpoint.

BSC adds new CCWs to be recognized by the 2703, SET MODE and POLL.
POLL is not used for point-to-point connections, so is a NOP.  SET
MODE selects the set of BSC protocol features used by the comms
device based on configuration tables in HASP and/or the cpu's OS.
This changes depending on whether the device is a 2770, 2780 or a
remote CPU acting as an RJE station, e.g. an 1130, a 360/20, etc.
N.B. Only the 2770/2780 protocol features have been tested.  There
are further, unimplemented, CCWs used to implement the BSC protocol
beyond point-to-point use, notably SEARCH.

An IBM 2780 emulator is available to connect to the BSC port and
communicate using the bisync protocol with HASP running on the
host system.  It may be found at

https://github.com/ghfbsd/rje-2780

Though it is agnostic as to the OS that hosts HASP, the 2780 RJE
program has only been tested with MTS as the host OS and may need
further work to function properly with other IBM OSes.  Carriage
control tapes, in particular, will be an area of divergence with
MTS behavior.